pyproject.toml
[tool.poetry]
name = "pureml"
version = "0.1.0"
description = ""
authors = ["vamsidhar muthireddy <vamsi.muthireddy@gmail.com>"]
readme = "README.md"



[tool.poetry.dependencies]
python = "^3.10 <3.11"
PyYAML = "^6.0"
pydantic = "^1.9.1"
numpy = "^1.23.1"
typer = {extras = ["all"], version = "^0.6.1"}
PyJWT = "^2.4.0"
requests = "^2.28.1"
python-dotenv = "^0.20.0"


[tool.poetry.dev-dependencies]
ipykernel = "^6.15.1"
lightgbm = "^3.3.2"
xgboost = "^1.6.1"
catboost = "^1.0.6"
optuna = "^2.10.1"
mlflow = "^1.28.0"
wandb = "^0.13.2"
pandas = "^1.4.3"
gradio = "^3.1.2"
scikit-learn = "^1.1.1"
fastparquet = "^0.8.1"
pyarrow = "^8.0.0"
pytest = "^7.1.3"
scipy = "^1.9.1"
torch = "^1.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pure = "pureml.cli.main:app"
